temporary commit; added assimp files to repo, improved animation loading, removed old dae loading from Karl (ThinMatrix), removed jGLTF physics is currently not working

+++ src/me/anno/fonts/signeddistfields/TextSDFGroup.kt
@@ -17,7 +17,7 @@ class TextSDFGroup(
     font: Font, text: String,
     charSpacing: Float
 ) : TextGroup(
-    font, text, charSpacing
+    font, text, charSpacing.toDouble()
 ) {
 
     // are draw-calls always expensive??
